+++ c/iotdm_client/lwm2m_objects/lwm2mserver_object.h
@@ -38,9 +38,9 @@ typedef struct TAG_object_lwm2mserver {
     uint16_t instanceId;    // Must be first member of structure
     IOTHUB_CHANNEL_HANDLE channelHandle;
     char resourceUpdated[LWM2MSERVER_PROPERTY_COUNT];
-    int propval_lwm2mserver_lifetime;        // /1/x/1
-    int propval_lwm2mserver_defaultminimumperiod;        // /1/x/2
-    int propval_lwm2mserver_defaultmaximumperiod;        // /1/x/3
+    int64_t propval_lwm2mserver_lifetime;        // /1/x/1
+    int64_t propval_lwm2mserver_defaultminimumperiod;        // /1/x/2
+    int64_t propval_lwm2mserver_defaultmaximumperiod;        // /1/x/3
     LWM2MSERVER_RESOURCE_CALLBACK lwm2mserver_lifetime_read_callback;
     LWM2MSERVER_RESOURCE_CALLBACK lwm2mserver_lifetime_write_callback;
     LWM2MSERVER_RESOURCE_CALLBACK lwm2mserver_defaultminimumperiod_read_callback;
@@ -53,9 +53,9 @@ IOTHUB_CLIENT_RESULT create_lwm2mserver_object(IOTHUB_CLIENT_HANDLE h, uint16_t 
 object_lwm2mserver *get_lwm2mserver_object(uint16_t instanceId);
 
 // Property setters for lwm2mserver object
-IOTHUB_CLIENT_RESULT set_lwm2mserver_lifetime(uint16_t instanceId, int value);
-IOTHUB_CLIENT_RESULT set_lwm2mserver_defaultminimumperiod(uint16_t instanceId, int value);
-IOTHUB_CLIENT_RESULT set_lwm2mserver_defaultmaximumperiod(uint16_t instanceId, int value);
+IOTHUB_CLIENT_RESULT set_lwm2mserver_lifetime(uint16_t instanceId, int64_t value);
+IOTHUB_CLIENT_RESULT set_lwm2mserver_defaultminimumperiod(uint16_t instanceId, int64_t value);
+IOTHUB_CLIENT_RESULT set_lwm2mserver_defaultmaximumperiod(uint16_t instanceId, int64_t value);
 
 #ifdef __cplusplus
 }
